+++ src/song/mod.rs
@@ -6,8 +6,7 @@
 
 mod msd;
 
-use either::Either;
-use std::{fs::File, io, io::BufReader, iter, mem::MaybeUninit, path::Path};
+use std::{fs::File, io, io::BufReader, path::Path};
 
 #[derive(Debug)]
 pub enum Error {
@@ -19,24 +18,6 @@ pub enum Error {
 pub enum Panel {
     None,
     Step,
-}
-
-impl From<msd::Step> for [Panel; 4] {
-    fn from(msd_step: msd::Step) -> Self {
-        match msd_step {
-            msd::Step::None => [Panel::None; 4],
-            msd::Step::Left => [Panel::Step, Panel::None, Panel::None, Panel::None],
-            msd::Step::Down => [Panel::None, Panel::Step, Panel::None, Panel::None],
-            msd::Step::Up => [Panel::None, Panel::None, Panel::Step, Panel::None],
-            msd::Step::Right => [Panel::None, Panel::None, Panel::None, Panel::Step],
-            msd::Step::DownLeft => [Panel::Step, Panel::Step, Panel::None, Panel::None],
-            msd::Step::UpLeft => [Panel::Step, Panel::None, Panel::Step, Panel::None],
-            msd::Step::LeftRight => [Panel::Step, Panel::None, Panel::None, Panel::Step],
-            msd::Step::UpDown => [Panel::None, Panel::Step, Panel::Step, Panel::None],
-            msd::Step::DownRight => [Panel::None, Panel::Step, Panel::None, Panel::Step],
-            msd::Step::UpRight => [Panel::None, Panel::None, Panel::Step, Panel::Step],
-        }
-    }
 }
 
 #[derive(Debug)]
@@ -54,86 +35,6 @@ pub struct Step<const PANELS: usize> {
 #[derive(Debug)]
 pub struct Steps<const PANELS: usize> {
     steps: Vec<Step<PANELS>>,
-}
-
-impl From<msd::Steps> for Steps<4> {
-    fn from(msd_steps: msd::Steps) -> Self {
-        let mut steps = Vec::new();
-
-        for notes in msd_steps.notes {
-            match notes {
-                msd::Notes::Eighth(step) => {
-                    steps.push(Step {
-                        panels: step.into(),
-                        duration: Duration::Eighth,
-                    });
-                }
-            }
-        }
-
-        Steps { steps }
-    }
-}
-
-impl From<(msd::Steps, msd::Steps)> for Steps<8> {
-    fn from(msd_steps: (msd::Steps, msd::Steps)) -> Self {
-        let mut steps = Vec::new();
-
-        // Combines the notes from both step charts into a single iterator, accounting for differing length as well.
-        let notes_0_len = msd_steps.0.notes.len();
-        let notes_1_len = msd_steps.1.notes.len();
-        let notes_iter = if notes_0_len > notes_1_len {
-            Either::Left(Either::Left(
-                msd_steps.0.notes.into_iter().zip(
-                    msd_steps.1.notes.into_iter().chain(
-                        iter::repeat(msd::Notes::Eighth(msd::Step::None))
-                            .take(notes_0_len - notes_1_len),
-                    ),
-                ),
-            ))
-        } else if notes_0_len < notes_1_len {
-            Either::Left(Either::Right(
-                msd_steps
-                    .0
-                    .notes
-                    .into_iter()
-                    .chain(
-                        iter::repeat(msd::Notes::Eighth(msd::Step::None))
-                            .take(notes_0_len - notes_1_len),
-                    )
-                    .zip(msd_steps.1.notes.into_iter()),
-            ))
-        } else {
-            Either::Right(
-                msd_steps
-                    .0
-                    .notes
-                    .into_iter()
-                    .zip(msd_steps.1.notes.into_iter()),
-            )
-        };
-
-        for (left_notes, right_notes) in notes_iter {
-            match (left_notes, right_notes) {
-                (msd::Notes::Eighth(left_step), msd::Notes::Eighth(right_step)) => {
-                    steps.push(Step {
-                        panels: {
-                            let mut whole = MaybeUninit::uninit();
-                            let ptr = whole.as_mut_ptr() as *mut [Panel; 4];
-                            unsafe {
-                                ptr.write(left_step.into());
-                                ptr.add(1).write(right_step.into());
-                                whole.assume_init()
-                            }
-                        },
-                        duration: Duration::Eighth,
-                    })
-                }
-            }
-        }
-
-        Steps { steps }
-    }
 }
 
 #[derive(Debug)]
@@ -198,52 +99,6 @@ impl Song {
         ))
         .map_err(|error| Error::Deserialization(error))
         .map(|song| song.into())
-    }
-}
-
-impl From<msd::Song> for Song {
-    fn from(song: msd::Song) -> Self {
-        let mut charts = Vec::new();
-
-        for single in song.single {
-            charts.push(Chart {
-                difficulty: single.0.into(),
-                meter: single.1,
-                style: Style::Single(single.2.into()),
-            })
-        }
-
-        for double in song.double {
-            charts.push(Chart {
-                difficulty: double.0.into(),
-                meter: double.1,
-                style: Style::Double((double.2, double.3).into()),
-            })
-        }
-
-        for couple in song.couple {
-            charts.push(Chart {
-                difficulty: couple.0.into(),
-                meter: couple.1,
-                style: Style::Couple((couple.2, couple.3).into()),
-            })
-        }
-
-        Song {
-            title: song.title,
-            subtitle: None,
-            artist: song.artist,
-            credit: song.msd,
-
-            bpm: song.bpm,
-            offset: song.gap,
-
-            background_file: song.back,
-            music_preview_file: song.select,
-            music_file: song.bgm,
-
-            charts: charts,
-        }
     }
 }
